[svn-r3237] Applying OSU's patch for increasing the rank to 32-bits (see ticket #42).

+++ src/mpid/ch3/include/mpidpre.h
@@ -66,7 +66,7 @@ typedef unsigned long MPID_Seqnum_t;
 typedef struct MPIDI_Message_match
 {
     int32_t tag;
-    int16_t rank;
+    int32_t rank;
     MPIR_Context_id_t context_id;
 }
 MPIDI_Message_match;
